Design-matrix columns UInt32 (was UInt16: wrapped above 65,535 matched precursors per scan)
timsTOF MS1 slices at 50 ng exceed the cap at the scout's +/-100 ppm; Hs.colptr[col + 1] wrapped to index 0.
Widened id_to_col / col / finalize_column! / SparsePrecMap / Score! signatures; regression test with 70,000
precursors in one scan (fails with BoundsError on the old code).

+++ src/Routines/SearchDIA/SearchMethods/SearchTypes.jl
@@ -206,7 +206,7 @@ mutable struct SimpleLibrarySearch{I<:IsotopeSplineModel} <: SearchDataStructure
     mass_err_samples::Vector{MassErrSample}
 
     # Indexing and scoring
-    id_to_col::SparsePrecMap{UInt16}
+    id_to_col::SparsePrecMap{UInt32}
     iso_splines::I
     
     # PSM scoring.
